requirements.txt
google-genai>=1.0.0
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
matplotlib>=3.8.0
scipy>=1.11.0
python-dotenv>=1.0.0
pytest>=8.0.0
